refactor: static method를 확장 함수로 리팩토링
